+++ src/index.ts
@@ -4,7 +4,10 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
+import dispatchTo from '../test/dispatchTo';
+import { testRendererForApps } from '../test/testrenderer';
 import { hideDialog, showDialog } from './ErrorDialog/errorDialogSlice';
+import rootReducer from './rootReducer';
 
 const ErrorDialogActions = { hideDialog, showDialog };
 
@@ -69,3 +72,11 @@ export {
 } from './Device/deviceLister';
 export { default as sdfuOperations } from './Device/sdfuOperations';
 export { defaultInitPacket, HashType, FwType } from './Device/initPacket';
+
+export { default as describeError } from './logging/describeError';
+
+export const testUtils = {
+    dispatchTo,
+    render: testRendererForApps,
+    rootReducer,
+};